Moved the nearby airport business logic to a partial class.
Hiba Seraj worked hard (sweat, blood and tears were shed on this assignment <≧❂◡❂≦>) to move the function and almost broke the code because of Alex due to namespace inconsistencies.
