+++ MainPage.xaml
@@ -2,7 +2,8 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:model="clr-namespace:Lab6_Starter.Model"
-             x:DataType="model:BusinessLogic" 
+             xmlns:model1="using:FWAPPA.Model"
+             x:DataType="model1:BusinessLogic" 
              x:Class="Lab6_Starter.MainPage"
              Title="Airports">
