+++ Model/BusinessLogic.cs
@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Diagnostics;
-using Lab6_Starter;
+﻿using System.Collections.ObjectModel;
+using Lab6_Starter.Model;
 
-namespace Lab6_Starter.Model;
+namespace FWAPPA.Model;
 
-public class BusinessLogic : IBusinessLogic
+public partial class BusinessLogic : IBusinessLogic
 {
-    private const int EARTH_RADUIS_IN_METERS = 6378000;
-    private const double MILES_PER_METER = 0.00062137;
     
     const int BRONZE_LEVEL = 42;
     const int SILVER_LEVEL = 84;
@@ -183,47 +178,6 @@ public class BusinessLogic : IBusinessLogic
         ObservableCollection<Weather> weathers = new ObservableCollection<Weather>();
         weathers.Add(new Weather("METAR KJFK 161853Z 21015G25KT 10SM -RA SCT020 BKN050", "TAF KJFK 161720Z 1618/1718 21015KT P6SM -RA BKN050"));
         return weathers;
-    }
-
-    /// <summary>
-    /// Find all airports within maxMiles of sourceAirport
-    /// </summary>
-    /// <param name="sourceAirport">The airport whose location to use as reference.</param>
-    /// <param name="maxMiles">How far the desired airports are to be from the sourceAirport.</param>
-    /// <returns></returns>
-    public ObservableCollection<Airport> CalculateNearbyAirports(Airport sourceAirport, int maxMiles)
-    {
-        ObservableCollection<Airport> nearbyAirports = [];
-        ObservableCollection<Airport> allAirports = GetAirports();
-
-
-        // Formula for when we are able to get coordinates
-        nearbyAirports.Add(new Airport("KFLD", "Fond du Lac", DateTime.Now, 1));
-        nearbyAirports.Add(new Airport("KMTW", "Manitowac", DateTime.Now, 1));
-        nearbyAirports.Add(new Airport("79C", "Brenner", DateTime.Now, 5));
-        nearbyAirports.Add(new Airport("KUNU", "Dodge County", DateTime.Now, 1));
-
-
-        // foreach (Airport destinationAirport in allAirports)
-        // {
-        //     // Haversine formula to find distance between two points
-        //     double sourceLatitudeRadians = sourceAirport.Latitude * (Math.PI / 180);
-        //     double destinationLatitudeRadians = destinationAirport.Latitude * (Math.PI / 180);
-        //     double latitudeDiffRadians = (destinationAirport.Latitude - sourceAirport.Latitude) * (Math.PI / 180);
-        //     double longitudeDiffRadians = (destinationAirport.Longitude - sourceAirport.Longitude) * (Math.PI / 180);
-        //     double flatDistance = Math.Pow(Math.Sin(latitudeDiffRadians / 2.0), 2.0) + 
-        //                (Math.Cos(sourceLatitudeRadians) * 
-        //                 Math.Cos(destinationLatitudeRadians) * 
-        //                 Math.Pow(Math.Sin(longitudeDiffRadians / 2.0), 2.0));
-        //     double angularDistance = 2 * Math.Atan2(Math.Sqrt(flatDistance), Math.Sqrt(1 - flatDistance));
-        //     double distanceInMeters = EARTH_RADUIS_IN_METERS * angularDistance;
-        //     double distanceInMiles = distanceInMeters * MILES_PER_METER;
-        //     if (distanceInMiles < maxMiles)
-        //     {
-        //         nearbyAirports.Add(destinationAirport);
-        //     }
-        // }
-        return nearbyAirports;
     }
 
     public Route GetRoute()
